bar is displayed
